Prelim NSLookup and RemoteFile
Have both nslookup and remotefile working in basic forms. NSLookup
requires some parsing and is utilizing PInvoke due to permissions to run
some code. Output is currently stored as a string and needs to be
parsed. RemoteFile is grabbing the remote file and saving it in the
users directory with the external filename.