Changing QueryPlugin Returns
Changed Query Plugins to return csv files. The main application now uses
GenericParser to parse the csv into a DataTable of which each plugin
type has one in a DataSet. A combobox is used to switch between which
table is to be viewed. The return type is still very volatile and can
change at any moment!